Revert the revert: go back to r19322, except for Version.java which we keep.
git-svn-id: file:///home/toad/git-migration/temprepository/trunk/freenet@19326 67a373e5-eb02-0410-a15c-ee090a768436

+++ src/freenet/support/io/PaddedEphemerallyEncryptedBucket.java
@@ -271,7 +271,7 @@ public class PaddedEphemerallyEncryptedBucket implements Bucket, SerializableToF
 		if(size < minPaddedSize) size = minPaddedSize;
 		if(size == minPaddedSize) return size;
 		long min = minPaddedSize;
-		long max = minPaddedSize << 1;
+		long max = (long)minPaddedSize << 1;
 		while(true) {
 			if(max < 0)
 				throw new Error("Impossible size: "+size+" - min="+min+", max="+max);
